+++ CHANGELOG.md
@@ -106,6 +106,7 @@ There are no breaking API changes in this release, nor any changes to how cluste
 - [PR #2602](https://github.com/rqlite/rqlite/pull/2602): Clearer DB extension driver name.
 - [PR #2603](https://github.com/rqlite/rqlite/pull/2603): Capture CRC32 of FullSink while data is being written to disk.
 - [PR #2605](https://github.com/rqlite/rqlite/pull/2605): Define Suffrage via Protobuf.
+- [PR #2606](https://github.com/rqlite/rqlite/pull/2606): Snapshot Restore checks CRCs.
 
 ## v9.4.5 (March 9th 2026)
 ### Implementation changes and bug fixes